requirements.txt
streamlit
docker
python-dotenv
openai
anthropic
requests
python-telegram-bot
discord.py
